Start moving some logic from being embedded in AV into the Rails Sprockets extensions
